Add more control for deciding whether to queue or drop
You can now drop a job submission based on the state of an existing job.
Indicate the job states for which you want to job the drop if it exists
in that state.

For example, if you pass nx: "ac", it will not add the job if a job
exists in redis in the active or complete states.