Properly requeue jobs for resubmissions of IDs in active state
We want the property that only one instance of a job with a given ID is
executing at a certain time.

Before this commit, if a job was submitted while a job was in the active
state, it could be run simultaneously with the currently running one.

This prevents the case where two jobs with the same ID might be
interleaved.